+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -4,6 +4,37 @@ CHANGELOG
 3.3.0
 -----
 
+ * Deprecated support for implicitly parsing non-string mapping keys as strings. Mapping keys that are no strings will
+   lead to a `ParseException` in Symfony 4.0. Use the `PARSE_KEYS_AS_STRINGS` flag to opt-in for keys to be parsed as
+   strings.
+
+   Before:
+
+   ```php
+   $yaml = <<<YAML
+   null: null key
+   true: boolean true
+   1: integer key
+   2.0: float key
+   YAML;
+
+   Yaml::parse($yaml);
+   ```
+
+   After:
+
+   ```php
+
+   $yaml = <<<YAML
+   null: null key
+   true: boolean true
+   1: integer key
+   2.0: float key
+   YAML;
+
+   Yaml::parse($yaml, Yaml::PARSE_KEYS_AS_STRINGS);
+   ```
+
  * Omitted mapping values will be parsed as `null`.
 
  * Omitting the key of a mapping is deprecated and will throw a `ParseException` in Symfony 4.0.